PR comments + fix import issue
