+++ vllm/model_executor/models/mixtral.py
@@ -427,9 +427,6 @@ class MixtralForCausalLM(nn.Module, SupportsLoRA):
 
         params_dict = dict(self.named_parameters())
         for name, loaded_weight in weights:
-            if "weight_shape" in name:
-                continue
-
             if "rotary_emb.inv_freq" in name:
                 continue
